+++ imio/dms/mail/browser/iconified_category.py
@@ -4,13 +4,8 @@
 """
 from collective.iconifiedcategory.browser.actionview import ApprovedChangeView as BaseApprovedChangeView
 from collective.iconifiedcategory.browser.tabview import ApprovedColumn as BaseApprovedColumn
-from imio.dms.mail.utils import add_file_to_approval
-from imio.dms.mail.utils import approve_file
-from imio.dms.mail.utils import can_approve
-from imio.dms.mail.utils import get_approval_annot
-from imio.dms.mail.utils import is_file_approved
+from imio.dms.mail.adapters import OMApprovalAdapter
 from imio.dms.mail.utils import logger  # noqa F401
-from imio.dms.mail.utils import remove_file_from_approval
 from plone import api
 from zope.i18n import translate
 
@@ -43,7 +38,7 @@ class ApprovedColumn(BaseApprovedColumn):
     def __init__(self, context, request, table):
         super(ApprovedColumn, self).__init__(context, request, table)
         # self.context is the mail here
-        self.a_a = get_approval_annot(self.context)
+        self.approval = OMApprovalAdapter(self.context)
         self.msg = u""
         self.base_class = "iconified-action"
 
@@ -62,7 +57,7 @@ class ApprovedColumn(BaseApprovedColumn):
             # to-approve class is used when state is prior to to_approve
             if self.is_deactivated(content):
                 # TODO esign : add condition on mailing (cannot edit if mailing)
-                if not self.a_a["users"] or not editable:
+                if not self.approval.approvers or not editable:
                     self.msg = u"Deactivated for approval"
                     return " to-approve "
                 else:
@@ -84,24 +79,25 @@ class ApprovedColumn(BaseApprovedColumn):
         elif av.p_state == "to_approve":
             self.base_class = "iconified-action-approved"
             # current user can approve now
-            if can_approve(self.a_a, av.userid, av.uid):
-                if self.a_a["files"][content.UID]["nb"][self.a_a["approval"]]["status"] == "a":
+            if self.approval.can_approve(av.userid, av.uid):
+                if self.approval.is_approved(content.UID):
                     self.msg = u"Already approved (click to change)"
                     return " active{}".format(editable)
                 self.msg = u"Waiting for your approval (click to approve)"
                 return editable
             # current user cannot approve now but is an approver
-            elif av.userid in self.a_a["users"]:
+            elif av.userid in self.approval.approvers:
                 # approver must yet approve
-                if self.a_a["users"][av.userid]["order"] > self.a_a["approval"]:
+                approver_number = self.approval.number(av.userid)
+                if approver_number and approver_number > self.approval.number():
                     self.msg = u"Waiting for other approval before you can approve"
                     return " waiting"
         # after a first approval, we show a partially or totally approved icon even for a previously approver
         # if content["approved"]:  # all approved  metadata not updated in approve_file function
-        if is_file_approved(self.a_a, content.UID):  # all approved
+        if self.approval.is_file_approved(content.UID):  # all approved
             self.msg = u"Totally approved"
             return " totally-approved"
-        elif is_file_approved(self.a_a, content.UID, totally=False):
+        elif self.approval.is_file_approved(content.UID, totally=False):
             self.msg = u"Partially approved. Still waiting for other approval(s)"
             return " partially-approved"
         else:
@@ -115,7 +111,7 @@ class ApprovedColumn(BaseApprovedColumn):
         if av.p_state in ("to_print", "to_be_signed", "signed", "sent"):
             return "#"
         # when to_approve, only an approver can click on the icon
-        if av.p_state == "to_approve" and not can_approve(self.a_a, av.userid, av.uid):
+        if av.p_state == "to_approve" and not self.approval.can_approve(av.userid, av.uid):
             return "#"
         return '{url}/@@{action}'.format(
             url=content.getURL(),
@@ -140,7 +136,7 @@ class ApprovedChangeView(BaseApprovedChangeView):
         super(ApprovedChangeView, self).__init__(context, request)
         self.parent = self.context.__parent__
         self.p_state = api.content.get_state(self.parent)
-        self.a_a = get_approval_annot(self.parent)
+        self.approval = OMApprovalAdapter(self.parent)
         self.reload = False
         self.user_id = None
         self.uid = self.context.UID()
@@ -156,12 +152,12 @@ class ApprovedChangeView(BaseApprovedChangeView):
         """ """
         values = {}
         status = 0
-        # logger.info("Before annot change: %s", self.a_a)
+        # logger.info("Before annot change: %s", self.approval.annot)
         # logger.info("Before values change: %s", old_values)
         if self.p_state == "to_approve":
             # in to_approve state, only an approver can approve or not
-            if can_approve(self.a_a, self.userid, self.uid):
-                if self.a_a["files"][self.uid]["nb"][self.a_a["approval"]]["status"] == "a":
+            if self.approval.can_approve(self.userid, self.uid):
+                if self.approval.is_approved(self.uid):
                     status = 0
                     self.msg = u"Already approved (click to change)"
                     # TODO TO BE HANDLED
@@ -169,8 +165,12 @@ class ApprovedChangeView(BaseApprovedChangeView):
                     self.msg = u"Waiting for your approval (click to approve)"
                     # the status is changed (if totally approved) in sub method
                     # must we pass self to update self.msg: no need for now because we reload in all cases !!
-                    ret, self.reload = approve_file(self.a_a, self.parent, self.context, self.userid, values=values,
-                                                    transition="propose_to_be_signed")
+                    ret, self.reload = self.approval.approve_file(
+                        afile=self.context,
+                        userid=self.userid,
+                        values=values,
+                        transition="propose_to_be_signed",
+                    )
                     status = int(ret)
         elif self.p_state not in ("to_print", "to_be_signed", "signed", "sent"):
             # before to_approve state, we can only enable or disable to_approve
@@ -178,19 +178,19 @@ class ApprovedChangeView(BaseApprovedChangeView):
                 values['to_approve'] = True
                 values['approved'] = False
                 status = 1
-                add_file_to_approval(self.a_a, self.uid)
+                self.approval.add_file_to_approval(self.uid)
                 self.msg = u"Activated for approval (click to deactivate)"
             else:
                 values['to_approve'] = False
                 values['approved'] = False
                 status = 0
-                remove_file_from_approval(self.a_a, self.uid)
+                self.approval.add_file_to_approval(self.uid)
                 self.msg = u"Deactivated for approval (click to activate)"
             self.reload = False
         else:
             # cannot be in after to_approve state because get_url column method ?
             logger.warn("IN else of approved change view ???")
-        # logger.info("After annot change: %s, ", self.a_a)
+        # logger.info("After annot change: %s, ", self.approval.annot)
         # logger.info("After values change: %s, %s", status, values)
         return status, values