+++ CHANGES.rst
@@ -30,6 +30,8 @@ Changelog
   [chris-adam, sgeulette]
 - Refactored VersionsTable into an iconified Categorizedtable.
   [chris-adam, sgeulette]
+- Added eSign admin table to manage approvals.
+  [chris-adam]
 
 3.0 (2021-09-30)
 ----------------